+++ rebook-product/src/main/java/com/be/rebook/product/entity/Product.java
@@ -1,6 +1,6 @@
 package com.be.rebook.product.entity;
 
-import com.be.rebook.domain.members.entity.Members;
+import com.be.rebook.product.entity.Members;
 import com.be.rebook.product.dto.ProductRequestDTO;
 import com.be.rebook.product.type.ProductStatus;
 import com.be.rebook.global.config.BaseEntity;
